pyproject.toml
[project]
name = "nilmformer"
version = "0.1.0"
description = "A sequence-to-sequence Transformer-based solution that accounts for electricity consumption time series non-stationarity"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "einops>=0.8.0",
    "einx>=0.3.0",
    "matplotlib>=3.9.2",
    "numpy>=2.1.3",
    "pandas>=2.2.3",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.5.2",
    "torch>=2.5.1",
    "xformers>=0.0.28.post3",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "ruff>=0.9.1",
]
